hoshino/modules/DDbeifen/hoshinobot-plugin-ddcheck/requirements.txt
aiofiles>=0.8.0
httpx>=0.22.0
jinja2<3.1.0
loguru>=0.6.0
Pillow>=9.1.0
playwright>=1.20.1
